rendering module refactor
