+++ src/snowflake/cli/api/rendering/jinja.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+
 from __future__ import annotations
 
 from pathlib import Path
@@ -19,16 +20,10 @@ from textwrap import dedent
 from typing import Dict, Optional
 
 import jinja2
-from click import ClickException
 from jinja2 import Environment, StrictUndefined, loaders
-from snowflake.cli.api.cli_global_context import cli_context
 from snowflake.cli.api.secure_path import UNLIMITED, SecurePath
 
 CONTEXT_KEY = "ctx"
-_YML_TEMPLATE_START = "<%"
-_YML_TEMPLATE_END = "%>"
-_PROJECT_TEMPLATE_START = "<!"
-_PROJECT_TEMPLATE_END = "!>"
 
 
 def read_file_content(file_name: str):
@@ -58,7 +53,7 @@ def procedure_from_js_file(env: jinja2.Environment, file_name: str):
 _CUSTOM_FILTERS = [read_file_content, procedure_from_js_file]
 
 
-def _env_bootstrap(env: Environment) -> Environment:
+def env_bootstrap(env: Environment) -> Environment:
     for custom_filter in _CUSTOM_FILTERS:
         env.filters[custom_filter.__name__] = custom_filter
 
@@ -86,51 +81,6 @@ class IgnoreAttrEnvironment(Environment):
             return self.undefined(obj=obj, name=argument)
 
 
-def get_snowflake_cli_jinja_env() -> Environment:
-    _random_block = "___very___unique___block___to___disable___logic___blocks___"
-    return _env_bootstrap(
-        IgnoreAttrEnvironment(
-            loader=loaders.BaseLoader(),
-            keep_trailing_newline=True,
-            variable_start_string=_YML_TEMPLATE_START,
-            variable_end_string=_YML_TEMPLATE_END,
-            block_start_string=_random_block,
-            block_end_string=_random_block,
-            undefined=StrictUndefined,
-        )
-    )
-
-
-def get_sql_cli_jinja_env():
-    _random_block = "___very___unique___block___to___disable___logic___blocks___"
-    return _env_bootstrap(
-        IgnoreAttrEnvironment(
-            loader=loaders.BaseLoader(),
-            keep_trailing_newline=True,
-            variable_start_string="&{",
-            variable_end_string="}",
-            block_start_string=_random_block,
-            block_end_string=_random_block,
-            undefined=StrictUndefined,
-        )
-    )
-
-
-def get_template_cli_jinja_env(template_root: SecurePath) -> Environment:
-    _random_block = "___very___unique___block___to___disable___logic___blocks___"
-    return _env_bootstrap(
-        Environment(
-            loader=loaders.FileSystemLoader(searchpath=template_root.path),
-            keep_trailing_newline=True,
-            variable_start_string=_PROJECT_TEMPLATE_START,
-            variable_end_string=_PROJECT_TEMPLATE_END,
-            block_start_string=_random_block,
-            block_end_string=_random_block,
-            undefined=StrictUndefined,
-        )
-    )
-
-
 def jinja_render_from_file(
     template_path: Path, data: Dict, output_file_path: Optional[Path] = None
 ) -> Optional[str]:
@@ -145,7 +95,7 @@ def jinja_render_from_file(
     Returns:
         None if file path is provided, else returns the rendered string.
     """
-    env = _env_bootstrap(
+    env = env_bootstrap(
         IgnoreAttrEnvironment(
             loader=loaders.FileSystemLoader(template_path.parent),
             keep_trailing_newline=True,
@@ -159,15 +109,3 @@ def jinja_render_from_file(
         return None
     else:
         return rendered_result
-
-
-def snowflake_sql_jinja_render(content: str, data: Dict | None = None) -> str:
-    data = data or {}
-    if CONTEXT_KEY in data:
-        raise ClickException(
-            f"{CONTEXT_KEY} in user defined data. The `{CONTEXT_KEY}` variable is reserved for CLI usage."
-        )
-
-    context_data = cli_context.template_context
-    context_data.update(data)
-    return get_sql_cli_jinja_env().from_string(content).render(**context_data)
