+++ src/snowflake/cli/api/utils/definition_rendering.py
@@ -24,10 +24,13 @@ from snowflake.cli.api.project.schemas.project_definition import (
     ProjectDefinition,
     ProjectProperties,
 )
+from snowflake.cli.api.rendering.jinja import CONTEXT_KEY
+from snowflake.cli.api.rendering.snowflake_yml_templates import (
+    get_snowflake_cli_jinja_env,
+)
 from snowflake.cli.api.utils.dict_utils import traverse
 from snowflake.cli.api.utils.graph import Graph, Node
 from snowflake.cli.api.utils.models import ProjectEnvironment
-from snowflake.cli.api.utils.rendering import CONTEXT_KEY, get_snowflake_cli_jinja_env
 from snowflake.cli.api.utils.types import Context, Definition
