+++ src/snowflake/cli/plugins/sql/manager.py
@@ -22,9 +22,9 @@ from typing import Dict, Iterable, List, Tuple
 
 from click import ClickException, UsageError
 from jinja2 import UndefinedError
+from snowflake.cli.api.rendering.sql_templates import snowflake_sql_jinja_render
 from snowflake.cli.api.secure_path import UNLIMITED, SecurePath
 from snowflake.cli.api.sql_execution import SqlExecutionMixin, VerboseCursor
-from snowflake.cli.api.utils.rendering import snowflake_sql_jinja_render
 from snowflake.cli.plugins.sql.snowsql_templating import transpile_snowsql_templates
 from snowflake.connector.cursor import SnowflakeCursor
 from snowflake.connector.util_text import split_statements
